issue 4226: store final run performance metrics during run finalization step

+++ api/src/main/java/com/epam/pipeline/entity/run/PipelineRunPerformanceMetric.java
@@ -23,7 +23,7 @@ import lombok.Getter;
 @Builder
 public class PipelineRunPerformanceMetric {
     private final PipelineRunPerformanceMetricsType type;
-    private final int max;
-    private final int avg;
-    private final long capacity;
+    private final double max;
+    private final double avg;
+    private final double capacity;
 }
